fixing errors

+++ 0x15-file_io/0-read_textfile.c
@@ -9,7 +9,7 @@
 
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-	int op, re, wr;
+	ssize_t op, re, wr;
 
 	char *buff;
